Created FirebaseExtensions module with an InitializeFirebase helper method
