+++ 404-brewery-bracket/HelperMethods/FirebaseExtensions.ts
@@ -1,0 +1,24 @@
+import { initializeApp, FirebaseApp } from "firebase/app";
+import { getFirestore, Firestore } from "firebase/firestore";
+
+function InitializeFirebase(): [FirebaseApp, Firestore] {
+  const firebaseConfig = {
+    apiKey: process.env.FIREBASE_APIKEY,
+    authDomain: process.env.FIREBASE_AUTHDOMAIN,
+    projectId: process.env.FIREBASE_PROJECTID,
+    storageBucket: process.env.FIREBASE_STORAGEBUCKET,
+    messagingSenderId: process.env.FIREBASE_MESSAGINGSENDERID,
+    appId: process.env.FIREBASE_APPID,
+    measurementId: process.env.FIREBASE_MEASUREMENTID,
+  };
+
+  // Initialize Firebase
+  const app = initializeApp(firebaseConfig);
+
+  // Create Firestore instance
+  const firestore = getFirestore(app);
+
+  return [app, firestore];
+}
+
+export { InitializeFirebase };
